+++ src/graph/expression_operators.cpp
@@ -95,7 +95,7 @@ Expr swish(Expr a) {
 }
 
 Expr gelu(Expr a) {
-  return Expression<SwishNodeOp>(a, 1.702f);
+  return Expression<GeluNodeOp>(a);
 }
 
 Expr operator-(Expr a) {